Docker/Auto Scalling Group/requirements.txt
boto3==1.26.133
grpcio==1.54.2
grpcio-tools==1.54.2
pymongo==4.3.3
dash==2.9.3
pandas==2.0.1
python-dotenv==1.0.0